customer endpoint is restricted
