+++ pkg/platform/customer/service.go
@@ -1,18 +1,37 @@
 package customer
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
 
 	dolittleK8s "github.com/dolittle/platform-api/pkg/dolittle/k8s"
+	"github.com/dolittle/platform-api/pkg/platform"
 	jobK8s "github.com/dolittle/platform-api/pkg/platform/job/k8s"
 	"github.com/dolittle/platform-api/pkg/platform/storage"
 	"github.com/dolittle/platform-api/pkg/utils"
 	"github.com/google/uuid"
+	"github.com/thoas/go-funk"
+	rbacv1 "k8s.io/api/rbac/v1"
+	"k8s.io/apimachinery/pkg/api/equality"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 )
+
+type service struct {
+	k8sclient               kubernetes.Interface
+	storageRepo             storage.RepoCustomer
+	platformOperationsImage string
+	platformEnvironment     string
+}
+
+type HttpCustomersResponse []platform.Customer
+
+type HttpCustomerInput struct {
+	Name string `json:"name"`
+}
 
 func NewService(
 	k8sclient kubernetes.Interface,
@@ -29,9 +48,10 @@ func NewService(
 }
 
 func (s *service) Create(w http.ResponseWriter, r *http.Request) {
-	// TODO need to figure out how we might want to expose this
-	disabled := true
-	if disabled {
+	userID := r.Header.Get("User-ID")
+	hasAccess := s.hasAccess(userID)
+
+	if !hasAccess {
 		utils.RespondWithError(w, http.StatusForbidden, "You do not have access")
 		return
 	}
@@ -88,9 +108,10 @@ func (s *service) Create(w http.ResponseWriter, r *http.Request) {
 }
 
 func (s *service) GetAll(w http.ResponseWriter, r *http.Request) {
-	// TODO need to figure out how we might want to expose this
-	disabled := true
-	if disabled {
+	userID := r.Header.Get("User-ID")
+	hasAccess := s.hasAccess(userID)
+
+	if !hasAccess {
 		utils.RespondWithError(w, http.StatusForbidden, "You do not have access")
 		return
 	}
@@ -103,6 +124,22 @@ func (s *service) GetAll(w http.ResponseWriter, r *http.Request) {
 	utils.RespondWithJSON(w, http.StatusOK, customers)
 }
 
+// hasAccess poormans security to lock down the endpoints, based on if the userID is in the rolebinding
+// rolebinding is hardcoded to platform-admin
 func (s *service) hasAccess(userID string) bool {
-	return false
+	ctx := context.TODO()
+	client := s.k8sclient
+	namespace := "system-api"
+	roleBinding, _ := client.RbacV1().RoleBindings(namespace).Get(ctx, "platform-admin", metav1.GetOptions{})
+
+	access := funk.Contains(roleBinding.Subjects, func(subject rbacv1.Subject) bool {
+		want := rbacv1.Subject{
+			Kind:     "User",
+			APIGroup: "rbac.authorization.k8s.io",
+			Name:     userID,
+		}
+		return equality.Semantic.DeepDerivative(subject, want)
+	})
+
+	return access
 }
